06-Python/multimedia_encrypt/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
cryptography>=42.0.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# ── Agentic AI analysis pipeline ─────────────────────────────────────────────
langgraph>=1.0.0
langchain-core>=0.3.0
spacy>=3.7.0
transformers>=4.40.0
torch>=2.2.0
openai-whisper>=20231117
pytesseract>=0.3.10
Pillow>=10.0.0
ultralytics>=8.0.0
opencv-python>=4.9.0
moviepy>=2.0.0
numpy>=1.26.0
